Move contract related fns to the threshold-ts lib
