+++ src/utils/getContract.ts
@@ -1,19 +1,5 @@
-import { AddressZero } from "../web3/utils/address"
-export { getSigner, getContract } from "../threshold-ts/utils/contract"
-
-export function getContractAddressFromTruffleArtifact(
-  truffleArtifact: { networks: { [chainID: string]: { address: string } } },
-  chainID: string | undefined = undefined
-) {
-  const networks = Object.keys(truffleArtifact.networks) as Array<
-    keyof typeof truffleArtifact.networks
-  >
-
-  return networks && networks.length > 0
-    ? (
-        truffleArtifact.networks[chainID ? chainID : networks[0]] as {
-          address: string
-        }
-      ).address
-    : AddressZero
-}
+export {
+  getSigner,
+  getContract,
+  getContractAddressFromTruffleArtifact,
+} from "../threshold-ts/utils/contract"
