+++ src/web3/utils/events.ts
@@ -1,18 +1,1 @@
-import { Contract, Event } from "@ethersproject/contracts"
-import { BlockTag } from "@ethersproject/abstract-provider"
-
-interface Options {
-  fromBlock?: BlockTag
-  toBlock?: BlockTag
-  filterParams: any[]
-  eventName: string
-}
-
-export const getContractPastEvents = async (
-  contract: Contract,
-  options: Options
-): Promise<Array<Event>> => {
-  const filter = contract.filters[options.eventName](...options.filterParams)
-
-  return await contract.queryFilter(filter, options.fromBlock, options.toBlock)
-}
+export { getContractPastEvents } from "../../threshold-ts/utils"
